requirements.txt
pillow==12.0.0
texture2ddecoder==1.0.5
